Clarify name of tokens file across repository
