+++ README.md
@@ -122,12 +122,12 @@ This must not be done too often otherwise the rate limit will be exceeded, and F
 Files
 --
 ### .ini (generated)
-* `tokens.ini` - Access token data generated by OAuth 2.0 credentials. Used to retrieve data from Fitbit API.
+* `tokens-<client_id>.ini` - Access token data generated by OAuth 2.0 credentials. Used to retrieve data from Fitbit API. A new file is generated for each client ID.
 
 ### Python (internal)
 * `token_handler.py` - Reads and writes token `.ini` files
 * `setup_access.py` - Used by `auth.sh` to generate token `.ini` files
-* `get_data.py` - Uses `python-fitbit` Python library to get Fitbit API data, using credentials from `tokens.ini`. Prints out in JSON to be parsed by `node_helper.js`.
+* `get_data.py` - Uses `python-fitbit` Python library to get Fitbit API data, using credentials from `tokens-<client_id>.ini`. Client credentials and data to fetch is passed in via stdin. Prints out in JSON, to be parsed by `node_helper.js`.
 
 ### Javascript (used by MagicMirror)
 * `node_helper.js` - Calls `get_data.py`, passes it to `MMM-Fitbit2.js`.
